model-deployment/deployment_pipeline/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers>=4.40.0
tokenizers>=0.19.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pydantic>=2.5.0
streamlit>=1.32.0
PyPDF2>=3.0.0
python-docx>=1.1.0